requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
black==23.7.0
boto3==1.28.3
botocore==1.31.3
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.3
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-redis==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.6
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
prompt-toolkit==3.0.39
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
